LAPR4-141 #server #developing Method numberOfPendingActivities implemented on class JpaDataRepository

+++ project/base.core/src/main/java/eapli/base/net/activityflux/repositories/DataRepository.java
@@ -6,7 +6,7 @@ import eapli.framework.domain.repositories.DomainRepository;
 public interface DataRepository
         extends DomainRepository<Long, Service> {
 
-    public Long numberOfPendingActivities(Long oMecanographicNumber);
+    public int numberOfPendingActivities(Long oMecanographicNumber);
 
     public int numberOfExpiredActivities();
